requirements.txt
APScheduler==3.6.3
astroid==2.5.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
future==0.18.2
idna==2.10
isort==5.8.0
lazy-object-proxy==1.5.2
mccabe==0.6.1
oauthlib==3.1.0
psycopg2-binary==2.8.6
PySocks==1.7.1
python-telegram-bot==13.4.1
python-twitter==3.5
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
wrapt==1.12.1
